Refactor FTP code to bring together common bits
