+++ crmprtd/wmb/download.py
@@ -1,6 +1,4 @@
 # Standard module
-import sys
-import csv
 import os
 import logging
 import logging.config
@@ -8,7 +6,8 @@ import ftplib
 from tempfile import SpooledTemporaryFile
 
 # Local
-from crmprtd import retry
+from crmprtd.download import retry, ftp_connect
+from crmprtd.download import FTPReader
 
 # Installed libraries
 import yaml
@@ -34,7 +33,8 @@ def download(args):
 
     try:
         # Connect FTP server and retrieve file
-        ftpreader = ftp_connect(args.ftp_server, args.ftp_file, log, auth)
+        ftpreader = ftp_connect(WMBFTPReader, args.ftp_server, args.ftp_file,
+                                log, auth)
 
         with SpooledTemporaryFile(
                 max_size=int(os.environ.get('CRMPRTD_MAX_CACHE', 2**20)),
@@ -42,62 +42,29 @@ def download(args):
             def callback(line):
                 tempfile.write('{}\n'.format(line))
 
-            log.info("Downloading %s", ftpreader.filename)
-            ftpreader.connection.retrlines('RETR {}'
-                                           .format(ftpreader.filename),
-                                           callback)
+            for filename in ftpreader.filenames:
+                log.info("Downloading %s", filename)
+                ftpreader.connection.retrlines('RETR {}'
+                                               .format(filename),
+                                               callback)
 
             tempfile.seek(0)
             yield tempfile
 
-    except Exception:
-        log.critical("Unable to process ftp")
+    except Exception as e:
+        log.exception("Unable to process ftp")
 
 
-def ftp_connect(host, path, log, auth):
-    log.info('Fetching file from FTP')
-    log.info('Listing.', extra={'host': host, 'path': path})
-
-    try:
-        ftpreader = FTPReader(host, auth['u'],
-                              auth['p'], path, log)
-        log.info('Opened a connection to host', extra={'host': host})
-    except ftplib.all_errors as e:
-        log.critical('Unable to load data from ftp source', exc_info=True)
-        sys.exit(1)
-
-    return ftpreader
-
-
-class FTPReader(object):
+class WMBFTPReader(FTPReader):
     '''Glue between the FTP_TLS class methods (which are callback based)
        and the csv.DictReader class (which is iteration based)
     '''
 
     def __init__(self, host, user, password, filename, log=None):
-        self.filename = filename
+        self.filenames = [filename]
 
         @retry(ftplib.error_temp, tries=4, delay=3, backoff=2, logger=log)
         def ftp_connect_with_retry(host, user, password):
             return ftplib.FTP_TLS(host, user, password)
 
         self.connection = ftp_connect_with_retry(host, user, password)
-
-    def csv_reader(self):
-        # Just store the lines in memory
-        # It's non-ideal but neither classes support coroutine send/yield
-        lines = []
-
-        def callback(line):
-            lines.append(line)
-
-        self.connection.retrlines('RETR {}'.format(self.filename), callback)
-
-        r = csv.DictReader(lines)
-        return r
-
-    def __del__(self):
-        try:
-            self.connection.quit()
-        except Exception:
-            self.connection.close()
